+++ src/components/MobileView/Inside.module.css
