+++ src/components/MobileView/InsidePower.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import styles from './InsideFileRight.module.css'
-import SmallBox from './polaroid.jsx' 
+import styles from './Inside.module.css'
+import SmallBox from './Polaroid.jsx' 
 
 const FilesLeftTitle = () => {
   return (
@@ -11,11 +11,13 @@ const FilesLeftTitle = () => {
 
                 <div className={styles.filespageright2}>
                 <h1 classname={styles.invi}>&nbsp;</h1>
-                <h1 align="center">TITLE</h1>
+                <h1 align="center">POWER</h1>
                 <img className={styles.SpyGunLogo}  src="public\Sponsors\ClosedFile\007-Logo.png"  alt="007Logo"/>
                 <img className={styles.target}      src="public\Sponsors\ClosedFile\Crosshair.png"  alt="Crosshair"/>
+                <div className={styles.center}><SmallBox/><SmallBox/></div>
                 <div className={styles.center}><SmallBox/></div>
-
+                <div className={styles.center}><SmallBox/><SmallBox/></div>
+                
                 </div>
 
             </div>
