+++ src/components/SponsorsPageInside/InsideL.module.css
@@ -1,28 +1,27 @@
-.filespagebgleft {
+.filespagebgL{
     background-color: #9E9683;
-    margin: 6vw;
-    rotate: -6deg;
+    margin: -6.5vw;
+    rotate: 6deg;
     height: 80vh;
-    width: 26vw; 
+    width: 26vw;
     position: relative;
 }
 
-.filespageleft1{
+.filespageL1{
     background-color: #B4AA96;
-    rotate: 3deg;
+    rotate: -3deg;
     height: 80vh;
     width: 26vw;
     position: relative;
 }
 
-.filespageleft2{
+.filespageL2{
     background-color: #CBC0A9;
-    rotate: 3deg;
+    rotate: -3deg;
     height: 80vh;
     width: 26vw;
     position: relative;
 }
-
 .center{
     display: flex;
     justify-content: center;
@@ -33,22 +32,24 @@
     color: transparent; 
 }
 
-.handprint {
-    overflow: hidden;
-    position: absolute;
-    top: -5vw;
-    right: -5vw;
-    width: auto;
-    height: 15vw; 
-    opacity: 0.25; 
-}
-
-.handcuffs {
+.target {
     overflow: hidden;
     display: flex;
     position: absolute;
-    bottom: -5vw;
-    right: -5vw;
-    width: 15vw; 
-    height: auto;
+    top: 0;
+    width: 7.5vw;
+    right: 1.5vw;
+    height: auto; 
+    opacity: 0.4; 
+}
+
+.SpyGunLogo{
+    overflow: hidden;
+    display: flex;
+    position: absolute;
+    top: 0;
+    width: 7.5vw;
+    left: 1.5vw;
+    height: auto; 
+    opacity: 2; 
 }
