+++ src/components/SponsorsPageInside/InsideLA.jsx
@@ -1,15 +1,15 @@
 import React from 'react'
-import styles from './InsideFileRight.module.css'
-import SmallBox from './polaroid.jsx' 
+import styles from './InsideL.module.css'
+import SmallBox from './Polaroid.jsx' 
 
 const FilesLeftTitle = () => {
   return (
     <>
-        <div className={styles.filespagebgright}>
+        <div className={styles.filespagebgL}>
 
-            <div className={styles.filespageright1}>
+            <div className={styles.filespageL1}>
 
-                <div className={styles.filespageright2}>
+                <div className={styles.filespageL2}>
                 <h1 classname={styles.invi}>&nbsp;</h1>
                 <h1 align="center">ASSOCIATE</h1>
                 <img className={styles.SpyGunLogo}  src="public\Sponsors\ClosedFile\007-Logo.png"  alt="007Logo"/>
