+++ src/components/SponsorsPageInside/InsideR.module.css
@@ -1,27 +1,28 @@
-.filespagebgright{
+.filespagebgR {
     background-color: #9E9683;
-    margin: -6.5vw;
-    rotate: 6deg;
+    margin: 6vw;
+    rotate: -6deg;
     height: 80vh;
-    width: 26vw;
+    width: 26vw; 
     position: relative;
 }
 
-.filespageright1{
+.filespageR1{
     background-color: #B4AA96;
-    rotate: -3deg;
+    rotate: 3deg;
     height: 80vh;
     width: 26vw;
     position: relative;
 }
 
-.filespageright2{
+.filespageR2{
     background-color: #CBC0A9;
-    rotate: -3deg;
+    rotate: 3deg;
     height: 80vh;
     width: 26vw;
     position: relative;
 }
+
 .center{
     display: flex;
     justify-content: center;
@@ -32,24 +33,22 @@
     color: transparent; 
 }
 
-.target {
+.handprint {
     overflow: hidden;
-    display: flex;
     position: absolute;
-    top: 0;
-    width: 7.5vw;
-    right: 1.5vw;
-    height: auto; 
-    opacity: 0.4; 
+    top: -5vw;
+    right: -5vw;
+    width: auto;
+    height: 15vw; 
+    opacity: 0.25; 
 }
 
-.SpyGunLogo{
+.handcuffs {
     overflow: hidden;
     display: flex;
     position: absolute;
-    top: 0;
-    width: 7.5vw;
-    left: 1.5vw;
-    height: auto; 
-    opacity: 2; 
+    bottom: -5vw;
+    right: -5vw;
+    width: 15vw; 
+    height: auto;
 }
